feat: Finish split of JsonValue and JsonValueFactory
This decoupling gives us more flexibility, and using a struct to implement the
iterator trait allows us to have a slightly more ergonomic API that borrows the
parser during iteration.

```rs
let parser = StreamingParser::new(ParserOptions::default());
for event in self.parser.feed(buffer) {
    // process event
}
```

Benchmarks will tell if the overhead of threading through the factory is
acceptable. Ideally with the inlining, it should be zero-cost, but on paper
we're now passing through a pointer to a lot more functions, though that pointer
_is not used_ in many of the implementations.